+++ src/service_gen.rs
@@ -22,7 +22,7 @@ impl TwirpServiceGenerator {
     fn generate_type_aliases(&mut self, buf: &mut String) {
         buf.push_str(&format!(
             "\n\
-                pub type PTReq<I> = {0}::PTReq<I>;\n\
+                pub type ServiceRequest<I> = {0}::ServiceRequest<I>;\n\
                 pub type PTRes<O> = {0}::PTRes<O>;\n",
             self.prost_twirp_mod()
         ));
@@ -51,7 +51,7 @@ impl TwirpServiceGenerator {
         let input_type = format_ident!("{}", method.input_type);
         let output_type = format_ident!("{}", method.output_type);
         quote! {
-            fn #name(&self, i: #prost_twirp::PTReq<#input_type>)
+            fn #name(&self, i: #prost_twirp::ServiceRequest<#input_type>)
                 -> #prost_twirp::PTRes<#output_type>
         }
     }