+++ src/service_run.rs
@@ -16,9 +16,6 @@ use prost::{DecodeError, EncodeError, Message};
 pub type FutReq<T> =
     Box<dyn Future<Output = Result<ServiceRequest<T>, ProstTwirpError>> + Send + 'static>;
 
-/// The type of every service request
-pub type PTReq<I> = ServiceRequest<I>;
-
 /// The type of every service response
 pub type PTRes<O> =
     Pin<Box<dyn Future<Output = Result<ServiceResponse<O>, ProstTwirpError>> + Send + 'static>>;